+++ src/factory/organizationType.ts
@@ -2,8 +2,18 @@
  * 組織タイプ
  */
 export enum OrganizationType {
+    /**
+     * 販売組織
+     */
     Corporation = 'Corporation',
-    MovieTheater = 'MovieTheater',
+    // 不要なので廃止(2022-05-30~)
+    // MovieTheater = 'MovieTheater',
+    /**
+     * 顧客組織
+     */
     Organization = 'Organization',
+    /**
+     * プロジェクト
+     */
     Project = 'Project'
 }