+++ CHANGELOG.md
@@ -11,10 +11,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 転送取引のIRecipientを最適化
+- 取引のIAgentを最適化
 
 ### Deprecated
 
 ### Removed
+
+- OrganizationType.MovieTheaterを廃止
 
 ### Fixed